Fix JavaScript warnings
Properties in module that are accessed by other modules must be defined
by 'var' rather then 'const'

(gnome-shell:24506): Gjs-WARNING **: 08:46:07.163: Some code accessed the property 'WorkspaceIsolator' on the module 'workspaceIsolater'. That property was defined with 'let' or 'const' inside the module. This was previously supported, but is not correct according to the ES6 standard. Any symbols to be exported from a module must be defined with 'var'. The property access will work as previously for the time being, but please fix your code anyway.

(gnome-shell:24506): Gjs-WARNING **: 08:48:53.236: Some code accessed the property 'CONF_DIR' on the module 'fileUtils'. That property was defined with 'let' or 'const' inside the module. This was previously supported, but is not correct according to the ES6 standard. Any symbols to be exported from a module must be defined with 'var'. The property access will work as previously for the time being, but please fix your code anyway.

(gnome-shell:24506): Gjs-WARNING **: 08:48:53.240: Some code accessed the property 'sessionManager' on the module 'sessionManager'. That property was defined with 'let' or 'const' inside the module. This was previously supported, but is not correct according to the ES6 standard. Any symbols to be exported from a module must be defined with 'var'. The property access will work as previously for the time being, but please fix your code anyway.

(gnome-shell:24506): Gjs-WARNING **: 08:51:49.487: Some code accessed the property 'WorkspaceManager' on the module 'workspaceManager'. That property was defined with 'let' or 'const' inside the module. This was previously supported, but is not correct according to the ES6 standard. Any symbols to be exported from a module must be defined with 'var'. The property access will work as previously for the time being, but please fix your code anyway.

(gnome-shell:24506): Gjs-WARNING **: 09:10:25.844: Some code accessed the property 'ObjectInterfaceDialog' on the module 'uiUtils'. That property was defined with 'let' or 'const' inside the module. This was previously supported, but is not correct according to the ES6 standard. Any symbols to be exported from a module must be defined with 'var'. The property access will work as previously for the time being, but please fix your code anyway.